improve ElementPositionWriter and add possibility to set  format and precision of column value output

+++ src/Structure/ElementPositionWriter.h
@@ -9,7 +9,7 @@ public:
     ElementPositionWriter(const std::string& fname);
 
     void addRow(double spos,
-                const std::vector<int>& row,
+                const std::vector<double>& row,
                 const std::string& elements);
 
 private:
